+++ app/src/components/actionBar/actionBar.css
@@ -1,0 +1,29 @@
+.actionBar{
+    font-size: 15px;
+    position: fixed;
+    text-align: center;
+    top: 0;
+    margin-top: 35px;
+    background-color: #f4f4f4;
+    width: 100%;
+    right: 5px;
+}
+.filterContainer{
+    display: inline-block;
+    margin: 0px 30px;
+    margin-top: 1px;
+}
+.dropDownContainer {
+    display: block;
+}
+.clearContainer{
+    position: absolute;
+    top: 53%;
+    display: inline-block;
+    cursor: pointer;
+    color: #8c8c8c;
+    transition: all 0.1s ease-in-out;
+}
+.clearContainer:hover{
+    color: #00a0e4;
+}
