+++ app/src/components/main/main.css
@@ -1,0 +1,10 @@
+.content{
+    background: #f4f4f4;
+  }
+  .cardListContainer{
+    margin-top: 150px;
+    background-color: #f4f4f4;
+    height: 80%;
+    overflow: auto;
+    position: fixed;
+  }
