+++ app/src/components/sort/sort.css
@@ -1,0 +1,21 @@
+.sorting{
+    font-size: 16px;
+    position: relative;
+    display: block;
+    text-align: center;
+}
+.appBar .header{
+    font-size: 23px;
+}
+.sortBtn{
+margin: 1%;
+font-size: 16px;
+cursor: pointer;
+transition: all 0.1s ease-in-out;
+}
+.sortBtn:hover{
+    font-size: 17px;
+}
+.MuiInputBase-root {
+    font-size: 0.85rem  !important;
+}
