--- src/App.js
@@ -1,12 +1,0 @@
-import './App.css';
-import Main from './components/main/main.js';
-
-function App() {
-  return (
-    <div className="App">
-      <Main/>
-    </div>
-  );
-}
-
-export default App;
