--- src/components/card/card.css
@@ -1,89 +1,0 @@
-.card{
-    padding: 12px;
-    text-align: center;
-    border: 1px solid #eeeeeeab;
-    width: 19%;
-    display: inline-block;
-    margin: 16px;
-    box-shadow: 0 1px 0px 0 #00a0e4;
-    border-radius: 4px;
-    background-color: white;    
-    position: relative;;
-}
- .card .header{
-    display: flex;
-    margin: 4px;
-    margin-bottom: 4%;
-    text-align: left;
-    color: #00a0e4;
-}
-.patient_icon{
-    position: absolute;
-    right: 0%;
-    top: -0.5%;
-}
-.patient_name{
-    margin: initial;
-    color: black;
-}
-.patient_data{
-    font-size: 13px;
-    text-align: left;
-    width: 96%;
-    margin: auto;
-    padding: 2px 6px 0px 6px;
-}
-.basicData{
-    border-bottom: 1px solid #eee;
-    font-size: 13px;
-    padding-bottom: 2%;
-    margin-bottom: 2%;
-}
-.patient_data label{
-    display: block;
-    font-weight: 400;
-    margin-bottom: 4px;
-    font-size: 12px;
-}
-.field{
-    margin-left: 11px;
-    font-weight: 100;
-    font-size: 12px;
-}
-.patient_data .offers label{
-    display: inline-block;
-    width: 50%;
-    text-align: center;
-    margin-top: 2%;
-    font-size: 13px;
-}
-.offers .field{
-    margin: auto;
-    display: block;
-    font-weight: 400;
-    font-size: 14px;
-    margin-top: 1%;
-}
-.avarageTime{
-    margin-top: 3%;
-    margin-bottom: 0px;
-    text-align: center;
-    font-size: 14px;
-}
-.avarageTime .field{
-display: block;
-margin: auto;
-font-size: 14px;
-margin-top: 1%;
-font-weight: 400;
-color: #00a0e4;
-}
-.accepted{
-color: #4cd964;
-}
-.canceled{
-color:#ff2d55;
-}
-.MuiSvgIcon-root {
-font-size: 1.75rem !important;
-}
