--- src/components/filter/filter.css
@@ -1,9 +1,0 @@
-.MuiSlider-colorPrimary{
-color:#00a0e4  !important;
-}
-.MuiTypography-body1{
-    font-size: 0.85rem  !important;
-}
-.MuiTypography-gutterBottom {
-    margin-bottom: 0.1em !important;
-}
